update validate script

+++ mlprodict/cli/validate.py
@@ -2,6 +2,7 @@
 @file
 @brief Command line about validation of prediction runtime.
 """
+import os
 from logging import getLogger
 from pandas import DataFrame
 from ..onnxrt.validate import validate_operator_opsets, summary_report  # pylint: disable=E0402
@@ -11,7 +12,7 @@ def validate_runtime(verbose=1, opset_min=9, opset_max=11,
                      check_runtime=True, runtime='CPU', debug=False,
                      models=None, out_raw="onnx_runtime_raw.xlsx",
                      out_summary="onnx_runtime_summart.xlsx",
-                     fLOG=print):
+                     dump_folder=None, fLOG=print):
     """
     Walks through most of :epkg:`scikit-learn` operators
     or model or predictor or transformer, tries to convert
@@ -31,6 +32,8 @@ def validate_runtime(verbose=1, opset_min=9, opset_max=11,
     :param debug: stops whenever an exception is raised
     :param out_raw: output raw results into this file (excel format)
     :param out_summary: output an aggregated view into this file (excel format)
+    :param dump_folder: folder where to dump information (pickle)
+        in case of mismatch
     :param fLOG: logging function
 
     .. cmdref::
@@ -49,8 +52,13 @@ def validate_runtime(verbose=1, opset_min=9, opset_max=11,
     models = None if models in (None, "") else models.strip().split(',')
     logger = getLogger('skl2onnx')
     logger.disabled = True
+    if not dump_folder:
+        dump_folder = None
+    if not dump_folder and not os.path.exists(dump_folder):
+        raise FileNotFoundError(dump_folder)
     rows = validate_operator_opsets(verbose, models=models, fLOG=fLOG,
-                                    runtime=runtime, debug=debug)
+                                    runtime=runtime, debug=debug,
+                                    dump_folder=dump_folder)
     df = DataFrame(rows)
     df.to_excel(out_raw, index=False)
     piv = summary_report(df)
